refactor: 后端重复逻辑收敛、Main 拆分 hooks 与渲染/网络优化
- 收敛 jsonResponse×5、isAuthorized×7、keys.ts 重复的 sha256Hex/StoredApiKey/listStoredKeys 到 _apikey.ts
- WebDAV isAuthorized 补空凭据 fail-closed（对齐 verifyBasicAuth 语义）
- authorizeApiKey 并行 get；touchLastUsed 60s 节流，消除每请求一次 R2 put
- 升级 @cloudflare/workers-types 消 7 处 R2 include @ts-ignore；tsconfig 开启 noUnusedLocals/noUnusedParameters/noImplicitOverride/noFallthroughCasesInSwitch
- 前端 errorMessage() 收敛 37 处 (error as Error).message；裸 401 Response 统一 textResponse
- Main.tsx 1756→1057 行：PathBar 独立，抽 useFolderListing/useMultiSelect/useFolderCounts/useKeyboardShortcuts/usePasteUpload/useDragDropUpload/useUploadInputs
- FileGrid React.memo + Main 侧 props 稳定化；PreviewDialog/SitesView/ImagesView React.lazy 拆 chunk（主包 702K→684K + 4 chunks）
- 新增 POST /api/counts 批量子项计数，替代逐目录 PROPFIND 的 N+1；api-e2e 补 4 项断言
- 删除 web-vitals 死依赖；IMPOROEMENT 计划追加第四轮
