+++ functions/api/_zip.ts
@@ -14,7 +14,6 @@ export async function listAllObjects(
     const listing = await bucket.list({
       prefix,
       cursor,
-      // @ts-ignore `include` is supported by R2 but missing from this types version.
       include: ["httpMetadata"],
     });
     objects.push(...listing.objects);
