+++ src/app/__tests__/transferQueue.test.tsx
@@ -1,5 +1,4 @@
 import { act, renderHook } from "@testing-library/react";
-import React from "react";
 
 import { processTransferTask } from "../transfer";
 import {
